refactor(frontend): split client.ts into modular endpoints and types
Extracts the monolithic client.ts logic into feature-specific endpoints, centralizes core fetch wrappers in core.ts, and isolates interfaces in types.ts. client.ts is preserved as a barrel file to ensure full backwards compatibility with consuming files and tests.
